+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # GraphDB AWS Terraform Module Changelog
 
 ## 2.5.0
-
 * Added support for managing Route 53 hosted zones and DNS records via external_dns_records module.
+* Enhanced KMS key and SNS topic configuration for Route 53 / CloudWatch availability alarms in us-east-1 to support encrypted alert delivery
 
 ## 2.4.0
 * Added support for Transit Gateway Attachments, Subnets and TGW Route table rules.